requirements.txt
discord.py==2.1.0
fastapi==0.89.1
pydantic==1.10.4
python-dotenv==0.21.1
requests==2.28.2